+++ src/composables/i18n.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 import { useI18n } from 'vue-i18n'
 
 export const useTranslation = () => {