+++ src/helpers/filters.ts
@@ -14,18 +14,30 @@ export function filterItems<T extends Filterable> (items: T[], searchQuery: stri
     if (column.trim() === '') return items
 
     return items.filter(item => {
-      if (Object.hasOwnProperty.bind(item)(column) && item[column] !== null) {
-        return item[column].toString().toLowerCase().includes(query)
+      if (columnFilterable(item, column)) {
+        return filterColumn(item, column, query)
       }
     })
   } else {
     return items.filter(item => {
       return headerNames.some(headerName => {
-        if (Object.hasOwnProperty.bind(item)(headerName) && item[headerName] !== null) {
-          return item[headerName].toString().toLowerCase().includes(search)
+        if (columnFilterable(item, headerName)) {
+          return filterColumn(item, headerName, search)
         }
       })
     })
+  }
+}
+
+const columnFilterable = (item: any, headerName: string) =>{
+  return Object.hasOwnProperty.bind(item)(headerName) && item[headerName] !== null
+}
+
+const filterColumn = (item: any, headerName: string, search: string) => {
+  try {
+    return item[headerName].toString().toLowerCase().includes(search)
+  } catch(_e) {
+    return false
   }
 }