+++ src/types/shims.d.ts
@@ -10,7 +10,7 @@ declare module '*.svg' {
 
 declare module '*.vue' {
   import { DefineComponent } from 'vue'
-  const component: DefineComponent<{}, {}, any>
+  const component: DefineComponent<object, object, any>
   export default component
 }